+++ src/BooksScreen/BooksScreen.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import { View } from 'react-native';
 
 import BookList from './components/BookList/BookList';
-import styles from './styles/booksScreen';
-import { booksNavOptions } from './constants/navigationOptions';
+import styles from './styles';
+import { booksNavOptions } from './constants';
 
 class BooksScreen extends Component {
   static navigationOptions = booksNavOptions;
