+++ src/BooksScreen/components/BookItem/BookItem.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { View, Text, Image, TouchableOpacity } from 'react-native';
 import { withNavigation } from 'react-navigation';
 
-import * as screenNames from '../../constants/screenNames';
+import * as screenNames from '../../../screenNames';
 
 import styles from './styles';
