+++ src/BooksScreen/components/BookItem/styles.js
