+++ src/BooksScreen/components/BookList/BookList.js
@@ -3,12 +3,12 @@ import { Text, View, FlatList } from 'react-native';
 import { connect } from 'react-redux';
 import axios from 'axios';
 
-import * as actionCreators from '../../actions';
-import * as statusCodes from '../../reducers/constants';
-import bookImage from './img/descarga.jpeg';
+import * as actionCreators from '../../../actions';
+import * as statusCodes from '../../../reducers/constants';
 import BookItem from '../BookItem/BookItem';
-import { Spinner } from '../common';
+import { Spinner } from '../../../common/';
 
+import bookImage from './img/descarga.jpeg';
 import styles from './styles';
 
 class BookList extends Component {
