+++ src/Navigation.js
@@ -1,10 +1,10 @@
 import { createMaterialTopTabNavigator, createStackNavigator } from 'react-navigation';
 
-import * as screenNames from './constants/screenNames';
-import TodoScreen from './TodoScreen';
-import BooksScreen from './BooksScreen';
-import BookDetailScreen from './BookDetailScreen';
-import styles from './styles/navigation';
+import * as screenNames from './screenNames';
+import TodoScreen from './TodoScreen/TodoScreen';
+import BooksScreen from './BooksScreen/BooksScreen';
+import BookDetailScreen from './BookDetailScreen/BookDetailScreen';
+import styles from './styles';
 
 const TabNavigator = createMaterialTopTabNavigator(
   {
