+++ src/TodoScreen/constants.js
@@ -1,0 +1,3 @@
+export const todoNavOptions = {
+  title: "ToDo's List"
+};
