+++ src/common/Spinner/Spinner.js
