+++ src/common/Spinner/styles.js
