--- src/constants/navigationOptions.js
@@ -1,12 +1,0 @@
-/**
- * Navigation Options of all screens
- *
- */
-
-export const todoNavOptions = {
-  title: "ToDo's List"
-};
-
-export const booksNavOptions = {
-  title: 'Books'
-};
